+++ packages/vee-validate/src/useFieldState.ts
@@ -1,7 +1,6 @@
 import { computed, reactive, ref, Ref, unref, watch } from 'vue';
-import { FormContextKey } from './symbols';
-import { FieldMeta, FieldState, MaybeRef } from './types';
-import { getFromPath, injectWithSelf, isEqual } from './utils';
+import { FieldMeta, FieldState, MaybeRef, PrivateFormContext } from './types';
+import { getFromPath, isEqual } from './utils';
 
 export interface StateSetterInit<TValue = unknown> extends FieldState<TValue> {
   initialValue: TValue;
@@ -20,7 +19,7 @@ export interface FieldStateComposable<TValue = unknown> {
 
 export interface StateInit<TValue = unknown> {
   modelValue: MaybeRef<TValue>;
-  standalone: boolean;
+  form?: PrivateFormContext;
 }
 
 let ID_COUNTER = 0;
@@ -29,8 +28,8 @@ export function useFieldState<TValue = unknown>(
   path: MaybeRef<string>,
   init: Partial<StateInit<TValue>>
 ): FieldStateComposable<TValue> {
-  const { value, initialValue, setInitialValue } = _useFieldValue<TValue>(path, init.modelValue, !init.standalone);
-  const { errorMessage, errors, setErrors } = _useFieldErrors(path, !init.standalone);
+  const { value, initialValue, setInitialValue } = _useFieldValue<TValue>(path, init.modelValue, init.form);
+  const { errorMessage, errors, setErrors } = _useFieldErrors(path, init.form);
   const meta = _useFieldMeta(value, initialValue, errors);
   const id = ID_COUNTER >= Number.MAX_SAFE_INTEGER ? 0 : ++ID_COUNTER;
 
@@ -76,9 +75,8 @@ interface FieldValueComposable<TValue = unknown> {
 export function _useFieldValue<TValue = unknown>(
   path: MaybeRef<string>,
   modelValue?: MaybeRef<TValue>,
-  shouldInjectForm = true
+  form?: PrivateFormContext
 ): FieldValueComposable<TValue> {
-  const form = shouldInjectForm === true ? injectWithSelf(FormContextKey, undefined) : undefined;
   const modelRef = ref(unref(modelValue)) as Ref<TValue>;
 
   function resolveInitialValue() {
@@ -170,9 +168,7 @@ function _useFieldMeta<TValue>(
 /**
  * Creates the error message state for the field state
  */
-export function _useFieldErrors(path: MaybeRef<string>, shouldInjectForm: boolean) {
-  const form = shouldInjectForm ? injectWithSelf(FormContextKey, undefined) : undefined;
-
+export function _useFieldErrors(path: MaybeRef<string>, form?: PrivateFormContext) {
   function normalizeErrors(messages: string | string[] | null | undefined) {
     if (!messages) {
       return [];