+++ packages/vee-validate/tests/useForm.spec.ts
@@ -1,7 +1,7 @@
 import { FormContext, useField, useForm } from '@/vee-validate';
 import { mountWithHoc, setValue, flushPromises, runInSetup } from './helpers';
 import * as yup from 'yup';
-import { Ref } from 'vue';
+import { onMounted, Ref } from 'vue';
 
 describe('useForm()', () => {
   const REQUIRED_MESSAGE = 'Field is required';
@@ -507,5 +507,117 @@ describe('useForm()', () => {
     jest.advanceTimersByTime(200);
     await flushPromises();
     expect(error?.textContent).toBe('not b');
+  });
+
+  // #3862
+  test('exposes controlled only values', async () => {
+    const spy = jest.fn();
+    const initial = {
+      field: '111',
+      createdAt: Date.now(),
+    };
+    mountWithHoc({
+      setup() {
+        const { controlledValues, handleSubmit } = useForm({
+          initialValues: initial,
+        });
+
+        const onSubmit = handleSubmit(values => {
+          spy({ values, controlled: controlledValues.value });
+        });
+
+        useField('field');
+
+        onMounted(onSubmit);
+
+        return {};
+      },
+      template: `
+        <div></div>
+      `,
+    });
+
+    await flushPromises();
+    expect(spy).toHaveBeenCalledTimes(1);
+    expect(spy).toHaveBeenLastCalledWith(
+      expect.objectContaining({
+        values: initial,
+        controlled: { field: initial.field },
+      })
+    );
+  });
+
+  // #3862
+  test('exposes controlled only via submission handler withControlled', async () => {
+    const spy = jest.fn();
+    const initial = {
+      field: '111',
+      createdAt: Date.now(),
+    };
+    mountWithHoc({
+      setup() {
+        const { handleSubmit } = useForm({
+          initialValues: initial,
+        });
+
+        const onSubmit = handleSubmit.withControlled(values => {
+          spy({ values });
+        });
+
+        useField('field');
+
+        onMounted(onSubmit);
+
+        return {};
+      },
+      template: `
+        <div></div>
+      `,
+    });
+
+    await flushPromises();
+    expect(spy).toHaveBeenCalledTimes(1);
+    expect(spy).toHaveBeenLastCalledWith(
+      expect.objectContaining({
+        values: { field: initial.field },
+      })
+    );
+  });
+
+  test('useFieldModel marks the field as controlled', async () => {
+    const spy = jest.fn();
+    const initial = {
+      field: '111',
+      field2: '222',
+      createdAt: Date.now(),
+    };
+    mountWithHoc({
+      setup() {
+        const { handleSubmit, useFieldModel } = useForm({
+          initialValues: initial,
+        });
+
+        const onSubmit = handleSubmit.withControlled(values => {
+          spy({ values });
+        });
+
+        const fields = useFieldModel(['field', 'field2']);
+
+        onMounted(onSubmit);
+
+        return {};
+      },
+      template: `
+        <div></div>
+      `,
+    });
+
+    await flushPromises();
+    expect(spy).toHaveBeenCalledTimes(1);
+    expect(spy).toHaveBeenLastCalledWith(
+      expect.objectContaining({
+        values: { field: initial.field, field2: initial.field2 },
+      })
+    );
   });
 });